Fixed #519: Give more details in 'highlight bullets' tooltips
for Highlight Points:
To activate, select some text in bullet format in the current slide and
click on this button.

for Highlight Background:
To activate, select some text in bullet format in the current slide and
click on this button.

for Highlight Text:
To activate, select the text you want to highlight and click on this button.