Let PowerPoint determine ack image dimensions
The previous implementation hardcodes the
acknowledgment image's scaling, relative to
the slide's dimensions.

As this causes stretching in the widescreen
slide type, allowing PowerPoint to determine the
correct dimensions of the image ensures that the
image is in it's native resolution.

As such, the alignment of the image is done after
the image has been applied onto the slide.