Crop Lab: Merge new features from PowerPointLabs-AY1617 (#1158)
* Add DevTeam table

* Let PowerPoint determine ack image dimensions (#1073)

The previous implementation hardcodes the
acknowledgment image's scaling, relative to
the slide's dimensions.

As this causes stretching in the widescreen
slide type, allowing PowerPoint to determine the
correct dimensions of the image ensures that the
image is in it's native resolution.

As such, the alignment of the image is done after
the image has been applied onto the slide.

* Revert "Fix stretching of acknowledgment image in widescreen #1072" (#1076)

* Add area leads

* Add more details on team members

* Add coding standards

* Add crop lab functionality

* Add review changes

* Missed a file

* Separate Crop Lab error handling UI and Logic

* Move Crop Lab duplicated code into error handler

* Speed up CropOutPadding by loading into memory

* Move CropToShape, CropToSame, CropToSlide to action framework

* Remove unneeded files

* Fix failing tests

* Move crop lab to action framework

* Change UIControl to MessageService

* Fix issues with CroptoSame and CropToSlide