#1640 Positions Lab Bug (Distribute) (#1653)
* Solved bug for Distribute Center, Horizontal, Vertical

* Throw an exception to warn users of corrupted images

1. Method to solve for distribute center/vertical/horizontal does not work here
2. Manual cut/paste of images also do not solve the problem

Method works fine for most images, therefore should try to preserve as best as possible, instead of sacrificing possible functionality for specific cases. Basic workaround would be to catch the COMException and print an error message.

* Added error message upon catching exception

Similar to Distribute by grid, manual cut/paste does not work. Furthermore the original method already worked for most images an should not be tempered with to work for a small group of corrupted images.

* Improved error solution by first removing corrupted shapes

Tested for PPT2016

* Added a break to remove prevent multiple error message

* Check for and delete duplicated shapes

Still need to return a suitable error message to remove any excess duplicates

* remove unneeded wrapper method

* Handle additional duplicated shapes

Tested on both 2013 and 2016

* Made error message more explicit

* Added thorough explanation for try-catch

* Updated comments to follow coding standards

* Updated comments to explain the need to run twice

* refactored 'noOf' to 'numOf' for better clarity

* added breaks for for loop